% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{activity_aes}
\alias{activity_aes}
\title{Activity aesthetics mapping}
\usage{
activity_aes(
  color = activity_scale(),
  opacity = activity_scale(),
  attributes = list()
)
}
\arguments{
\item{color}{The scale used for the activity color,}

\item{opacity}{The scale used for the activity opacity.}

\item{attributes}{A list of additional (fixed - non changing) SVG attributes to be added to each activity.}
}
\value{
An aesthetics mapping for `animate_process`.
}
\description{
Activity aesthetics mapping
}
\examples{
data(example_log)

\donttest{
}

}
\seealso{
\code{\link{animate_process}}, \code{\link{activity_scale}}
}
