% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{activity_scale}
\alias{activity_scale}
\title{Activity scale mapping values to aesthetics}
\usage{
activity_scale(
  attribute = NULL,
  scale = c("identity", "linear", "sqrt", "log", "quantize", "ordinal", "time"),
  domain = NULL,
  range = NULL
)
}
\arguments{
\item{attribute}{This may be (1) the name of the event attribute to be used as values,
(2) a data frame with three columns (act, time, value) in which the values in the act column are matching the activity identifier of the supplied event log, or
(3) a constant value that does not change over time.}

\item{scale}{Which D3 scale function to be used out of `identity`, `linear`, `sqrt`, `log`, `quantize`, `ordinal`, or `time`.}

\item{domain}{The domain of the D3 scale function. Can be left NULL in which case it will be automatically determined based on the values.}

\item{range}{The range of the D3 scale function. Should be a vector of two or more numerical values.}
}
\value{
A scale to be used with `mapping_activity`
}
\description{
Creates a `list` of parameters suitable to be used as activity scale in (\code{\link{activity_aes}}) for mapping values to certain aesthetics of the activity in a process map animation.
 Refer to the d3-scale documentation (https://github.com/d3/d3-scale) for more information about how to set `domain` and `range` properly.
}
\examples{
data(example_log)


}
\seealso{
\code{\link{animate_process}}, \code{\link{activity_aes}}
}
