% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorations.R
\name{activity_select_decoration}
\alias{activity_select_decoration}
\title{Decoration callback for activity selection}
\usage{
activity_select_decoration(
  stroke_dasharray = "2",
  stroke_width = "2",
  stroke = "black"
)
}
\arguments{
\item{stroke_dasharray}{Sets the `stroke-dasharray` attribute for selected activities.}

\item{stroke_width}{Sets the `stroke-width` attribute for selected activities.}

\item{stroke}{Sets the `stroke` attribute for selected activities.}
}
\value{
A JavaScript callback function called when activity selection changes.
}
\description{
Decoration callback for activity selection
}
\examples{
# Create a decoration callback that increases the activity stroke width
activity_select_decoration(stroke_width = "5")

}
\seealso{
animate_process
}
