% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\name{attribution_osm}
\alias{attribution_osm}
\title{Standard attribution}
\usage{
attribution_osm()
}
\value{
The attribution character vector.
}
\description{
This is the standard attribution advised for OPenStreetMap tiles.
}
\examples{
attribution_osm()
}
