% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_log}
\alias{example_log}
\title{Example event log used in documentation}
\format{
An bupaR event log
}
\usage{
example_log
}
\description{
Example event log used in documentation
}
\keyword{datasets}
