% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{icon_marker}
\alias{icon_marker}
\title{Standard map marker}
\usage{
icon_marker()
}
\value{
SVG code for a map  marker.
}
\description{
The marker is based on Material Design (Apache 2.0 License): https://material.io/
}
\examples{
icon_marker()
}
