% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processanimateR.R
\name{processanimaterOutput}
\alias{processanimaterOutput}
\title{Create a process animation output element}
\usage{
processanimaterOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output variable to read the animation from}

\item{width, height}{Must be a valid CSS unit (like 100%, 400px, auto) or a number,
which will be coerced to a string and have px appended.}
}
\description{
Renders a renderProcessanimater within an application page.
}
