% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{processAnimaterProxy}
\alias{processAnimaterProxy}
\title{Create a processAnimater proxy}
\usage{
processAnimaterProxy(shinyId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{shinyId}{Widget unique id.}

\item{session}{Shiny session.}
}
\description{
Avoids to re-render the same graph. Must be called on the server side
}
