% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processanimateR.R
\name{renderProcessanimater}
\alias{renderProcessanimater}
\title{Renders process animation output}
\usage{
renderProcessanimater(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{The expression generating a process animation (animate_process).}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.}
}
\description{
Renders a SVG process animation suitable to be used by processanimaterOutput.
}
