% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorations.R
\name{token_select_decoration}
\alias{token_select_decoration}
\title{Decoration callback for token selection}
\usage{
token_select_decoration(stroke = "black")
}
\arguments{
\item{stroke}{Sets the `stroke` attribute of selected tokens.}
}
\value{
A JavaScript callback function called when token selection changes.
}
\description{
Decoration callback for token selection
}
\examples{
# Create a decoration callback that paints tokens red
token_select_decoration("red")

}
\seealso{
animate_process
}
