% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{updateTimeline}
\alias{updateTimeline}
\title{Update processAnimater timeline}
\usage{
updateTimeline(proxy, duration)
}
\arguments{
\item{proxy}{Proxy id.}

\item{duration}{New duration.}
}
\description{
Update processAnimater timeline
}
