% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{activity_aes}
\alias{activity_aes}
\title{Activity aesthetics mapping}
\usage{
activity_aes(
  color = activity_scale(),
  linecolor = activity_scale(),
  opacity = activity_scale(),
  attributes = list()
)
}
\arguments{
\item{color}{The scale used for the activity color,}

\item{linecolor}{The scale used for the activity color,}

\item{opacity}{The scale used for the activity opacity.}

\item{attributes}{A list of additional (fixed - non changing) SVG attributes to be added to each activity.}
}
\value{
An aesthetics mapping for `animate_process`.
}
\description{
Activity aesthetics mapping
}
\examples{
data(example_log)

animate_process(patients,
                mode = "absolute",
                legend = "act_color",
                duration = 300,
                mapping_activity = activity_aes(color = activity_scale("employee", scale = "ordinal", range = c("red", "green", "blue", "yellow")),
                                                linecolor = activity_scale("time", scale = "time", range = c("white", "black")),
                                                opacity = activity_scale("time", scale = "time", range = c(0.3, 1.0))))


}
\seealso{
\code{\link{animate_process}}, \code{\link{activity_scale}}
}
