% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\name{attribution_osm}
\alias{attribution_osm}
\title{Standard attribution advised for OpenStreetMap tiles.}
\usage{
attribution_osm()
}
\value{
The attribution character vector.
}
\description{
Standard attribution advised for OpenStreetMap tiles.
}
\examples{
attribution_osm()
}
