% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\name{icon_circle}
\alias{icon_circle}
\title{Standard circle marker}
\usage{
icon_circle()
}
\value{
SVG code for a map  marker.
}
\description{
The marker is based on Material Design (Apache 2.0 License): https://material.io/
}
\examples{
icon_circle()
}
