% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processanimateR.R
\name{processanimaterOutput}
\alias{processanimaterOutput}
\title{Create a process animation output element}
\usage{
processanimaterOutput(outputId, width = NULL, height = NULL)
}
\arguments{
\item{outputId}{output variable to read the animation from}

\item{width}{The desired width of the SVG}

\item{height}{The desired height of the SVG}
}
\description{
Renders a renderProcessanimater within an application page
}
