% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\name{renderer_graphviz}
\alias{renderer_graphviz}
\title{Render as a plain graph}
\usage{
renderer_graphviz()
}
\value{
A rendering function to be used with \code{\link{animate_process}}
}
\description{
Render as a plain graph
}
\examples{
TODO
}
