% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\name{renderer_leaflet}
\alias{renderer_leaflet}
\title{Render as graph on a geographical map}
\usage{
renderer_leaflet(node_coordinates, layer = c(paste0("new L.TileLayer('",
  tile, "',"), paste0("{ attribution : '", attribution_osm(), "'})")),
  tile = "http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
  options = list(), grayscale = TRUE, act_icon = icon_marker(),
  start_icon = icon_circle(), end_icon = icon_circle())
}
\arguments{
\item{node_coordinates}{A data frame with node coordinates in the format `act`, `lat`, `lng`.}

\item{layer}{The JavaScript code used to create a Leaflet layer. A TileLayer is used as default value.}

\item{tile}{The URL to be used for the standard Leaflet TileLayer.}

\item{options}{A named list of leaflet options, such as the center point of the map and the initial zoom level.}

\item{grayscale}{Whether to apply a grayscale filter to the map.}

\item{act_icon}{The SVG code used for the activity icon.}

\item{start_icon}{The SVG code used for the start icon.}

\item{end_icon}{The SVG code used for the end icon.}
}
\value{
A rendering function to be used with \code{\link{animate_process}}
}
\description{
Render as graph on a geographical map
}
\examples{
data(example_log)

# Animate the example process with activities placed in some locations
animate_process(example_log,
  renderer = renderer_leaflet(
    node_coordinates = data.frame(
       act = c("A", "B", "C", "D", "Start", "End"),
       lat = c(63.443680, 63.426925, 63.409207, 63.422336, 63.450950, 63.419706),
       lng = c(10.383625, 10.396972, 10.406418, 10.432119, 10.383368, 10.252347),
       stringsAsFactors = FALSE),
      options = list(center = c(63.412273, 10.399590), zoom = 12)),
  duration = 5, repeat_count = Inf)

}
\seealso{
animate_process
}
