% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\name{renderer_leaflet}
\alias{renderer_leaflet}
\title{Render as graph on a geographical map}
\usage{
renderer_leaflet(node_coordinates, edge_coordinates = NULL,
  tile = "http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
  center = c(37.8, -96.9), zoom = 4)
}
\arguments{
\item{node_coordinates}{a data frame with node coordinates in the format `node`, `lat`, `lng`}

\item{edge_coordinates}{an (optional) data frame with edge path coordinates format TBD}
}
\value{
A rendering function to be used with \code{\link{animate_process}}
}
\description{
Render as graph on a geographical map
}
\examples{
TODO
}
